Delaunay triangulation using Bowyer-Watson algorithm
The TriangleMesh struct can be initialized with a list of points and will generate the indices for a Delaunay triangulation of those points using the Bowyer-Watson algorithm.

The triangulation is accessible as a list of indices for all triangles that reference the original points in the order they were provided.

The implementation is based on the pseudo code found at the following link: https://en.wikipedia.org/wiki/Bowyer–Watson_algorithm."